feat(Volt): AIOpMode
