+++ Volt/src/main/kotlin/dev/kingssack/volt/ai/AITool.kt
@@ -1,0 +1,3 @@
+package dev.kingssack.volt.ai
+
+data class AITool(val name: String, val description: String, val inputSchema: Map<String, Any>)
