+++ Volt/src/main/kotlin/dev/kingssack/volt/ai/ActionDescriptor.kt
@@ -1,15 +1,5 @@
 package dev.kingssack.volt.ai
 
-data class ActionParameter(
-    val name: String,
-    val type: String,
-    val description: String,
-    val required: Boolean = true,
-    val defaultValue: Any? = null,
-    val min: Double? = null,
-    val max: Double? = null,
-)
-
 data class ActionDescriptor(
     val id: String,
     val name: String,
@@ -18,5 +8,3 @@ data class ActionDescriptor(
     val parameters: List<ActionParameter>,
     val returns: String = "Action",
 )
-
-data class AITool(val name: String, val description: String, val inputSchema: Map<String, Any>)
