+++ Volt/src/main/kotlin/dev/kingssack/volt/ai/ActionParameter.kt
@@ -1,0 +1,11 @@
+package dev.kingssack.volt.ai
+
+data class ActionParameter(
+    val name: String,
+    val type: String,
+    val description: String,
+    val required: Boolean = true,
+    val defaultValue: Any? = null,
+    val min: Double? = null,
+    val max: Double? = null,
+)
